requirements.txt
aminodorksfix
python-telegram-bot
httpx